+++ index.js
@@ -6,6 +6,5 @@ module.exports = {
   included: function(app) {
     this._super.included(app);
     this.app.import(app.bowerDirectory + '/firebase/firebase.js');
-    this.app.import(app.bowerDirectory + '/ember-inflector/ember-inflector.js');
   }
 };